feat: add first version of types to burger-elements component
